s2: Allow folio_daemon to write to /dev/uinput
Commit aosp/836840 changes permissions for accessing /dev/uinput. That
broke the folio_daemon.

This daemon is responsible for supportic magnetic folio cases on Pixel
2. We need to continue to support this operation.

Allow the folio daemon to access /dev/uinput by enrolling it into the
uhid group.

Note: this is not an sepolicy issue, but a permission issue.

Bug: 120306105
Test: adb logcat | grep uinput
Before the change, see a bunch of permission denials from folio_daemon.
After the change, not seeing anything in the logcat.

Change-Id: Idb33c8427d55b165f5aca2a68c09fa7e749fef57