s2: Move ACDB blobs to system
In our libacdbloader.so, the path "system" is specified, and acdb blobs were moved to "vendor",
apparently because of this the interlocutor hears an echo of his voice, so for now we return them back.

This solves the echo problem on the interlocutor's side and fixes volume control during a call.

This reverts commit 8dbcc749dcceae2baebfeb530ac1c1aa75c9cf77.

Change-Id: I51def61c634fd583fd3afed127028c082badc17f